fix: centralizando funções do crud na entidade tag
